added the 3rd version of the SAW generator
this version uses the Ariadne thread method just like the very first version, but the code runs way faster than the first version. Decorator is used to keep track of time in the program, it is not vital to the creation of the self avoiding walk but is always nice to have.